Add createResponseFuture for its handler. Remove oplockLevel in DiskEntry. Add more interface for create in DiskShare.
Remove the oplockLevel in DiskEntry. Removing the openedFileMap for the oplockLevel tracking. The client must track its oplockLevel by itself. Add more interface for create in DiskShare. It's not going to work for oplock implement if we didn't got interface providing oplockLevel information in CreateResponse. Add passing a Future of a createResponse to the createResponseHandler. Added FutureWrapper to make sure type consist. Public the constructor of File and Directory to ensure client able to create DiskEntry instance.